requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-paypal==1.0.0
djangorestframework==3.11.0
flower==0.9.5
humanize==2.5.0
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
Pillow==7.2.0
prometheus-client==0.8.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
